requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.17
python-dotenv>=1.0.0
mutagen>=1.47.0
requests>=2.32.0
musicbrainzngs>=0.7.1
beautifulsoup4>=4.12.0
Pillow>=10.4.0